GUI TURTLE 1.5
Added new button on the main screen. It leads to a new window with two buttons. One is custom shape and the other one is letter. SHAPE allows you to create your own shape. And the LETTER allows you to draw a letter you want. THE LETTER FUNCTION IS NOT FINISHED YET!